Add multiprocess wrapper.
The gist: This changes how tasks are run. They're run in a secondary child process controlled by multiprocess.Pool. The subprocess is not respawned for every task - it is long-lived. The subprocess will be respawned only in the event of a timeout or if its memory consumption grows too high.

This is implemented with a new setting, `MODE`, which can be either 'single' or 'multi'. It defaults to 'single', i.e., without this change.

Relevant tidbits:

The subprocess is created with a fork().
* All memory is copy-on-write, so it is not 2x.
* The parent process never actually runs tasks: because of this, it doesn't have any memory dedicated to lazily loaded connections or data. This has the side effect of every spawned subprocess needing to reload all lazy content.
* A new GIL is present, allowing termination without fear of hanging the parent process.
* Exceptions and exception handling should function identically - exceptions and stack traces are pickled and sent up to the parent. The only difference will be the base line of the stack trace. tblib is used for this pickling.

The inspiration for this change is primarily to deal with the hanging / timeouts we've seen when running within Docker. There are other side effects as well.

Benefits:
* Hanging tasks can no longer ignore the given timeout - they end immediately upon timeout, even if they're in a syscall from a C library.
* Respawning the subprocess is significantly faster than exiting and restarting the entire process.

Drawbacks:
* Respawning the subprocess takes about a tenth of a second. That's not bad on its own, but lazily loaded database connections can increase this to around three seconds of setup. That gets added to a tasks' latency.