Handle some SQS throttling more gracefully.
Checking queue size seems to be throttled more aggressively than
other operations. Because we only use this check to prioritize
queues to pull messages from, assuming the queue has messages
should be a good strategy when we encounter this.

When a queue doesn't have messages, we will poll it for a few seconds
and move on to the next one.

I introduced a "queue needs work" concept so that we still have a
"queue is empty" concept that isn't fuzzy.