Change shutdown log statement to be a warning.
Since Kale is meant to run infinitely any shutdown is technically abnormal. A warning here is a perfect example of logging something that is abnormal but does not cause an error in the code. Warning level log statement will help track down issues when the engine shuts down in the middle of working on a single task.