Worker checks Task should_run_task to decide if it should execute run_task or republish with delay (#60)
Worker now allows Tasks to override should_run_task. The default behavior is to return True and execute run_task immediately after should_run_task. If a Task returns False, Worker will not execute run_task and republish the message with a delay. This allows API users to cancel the execution of a task without forcing them to raise an exception.